requirements.txt
flask
pymongo
gunicorn
mysqlclient
cassandra-driver
scales
